refactor: migrate to pnpm and upgrade all packages to latest
- Migrate from npm to pnpm (package-lock.json → pnpm-lock.yaml)
- Upgrade React 18 → 19, Next.js 15 → 16, and all other dependencies
- Fix hydration mismatch by replacing dynamic imports with static imports
- Update next-seo v7 API (NextSeo → generateNextSeo, DefaultSeo → generateDefaultSeo)
- Add Turbopack config and migrate images.domains to images.remotePatterns
- Separate browser-safe OGP utilities from server-only jsdom code
- Remove unused usePostList.ts hook
- Add renovate.json with stabilityDays: 1 for update delay
