+++ .npmrc
@@ -1,0 +1,1 @@
+onlyBuiltDependencies=skia-canvas,esbuild,sharp
