+++ netlify.toml
@@ -1,6 +1,6 @@
 [build]
     publish = "out"
-    command = "npm run build"
+    command = "pnpm run build"
 
 [build.environment]
     NETLIFY_NEXT_PLUGIN_SKIP = "true"
