+++ package.json
@@ -1,7 +1,11 @@
 {
   "private": true,
+  "packageManager": "pnpm@10.17.0",
   "engines": {
     "node": ">=18"
+  },
+  "pnpm": {
+    "onlyBuiltDependencies": ["skia-canvas", "esbuild", "sharp"]
   },
   "scripts": {
     "predev": "node scripts/convert-toml-to-json.js",
@@ -9,79 +13,80 @@
     "prebuild": "node scripts/convert-toml-to-json.js",
     "generate-og": "node scripts/generate-og-images.js || ./scripts/generate-og-images.sh",
     "generate-og-shell": "bash ./scripts/generate-og-images.sh",
-    "build": "node scripts/fetch-github-profile.js && npm run generate-og && next build && echo 'Build completed'",
-    "start": "npx serve@latest out",
+    "build": "node scripts/fetch-github-profile.js && pnpm run generate-og && next build && echo 'Build completed'",
+    "start": "pnpm dlx serve@latest out",
     "postbuild": "next-sitemap",
     "analyze": "ANALYZE=true next build",
     "test": "jest",
     "test:watch": "jest --watch"
   },
   "dependencies": {
-    "@chakra-ui/icons": "^1.1.1",
-    "@chakra-ui/react": "^2.8.2",
+    "@chakra-ui/icons": "^2.2.4",
+    "@chakra-ui/react": "^3.30.0",
     "@chakra-ui/theme": "^3.4.6",
     "@chakra-ui/theme-utils": "^2.0.21",
     "@emotion/react": "^11.14.0",
-    "@emotion/styled": "^11.14.0",
+    "@emotion/styled": "^11.14.1",
     "@iarna/toml": "^2.2.5",
     "@iconify/icons-simple-icons": "^1.2.74",
-    "@iconify/react": "^5.2.1",
+    "@iconify/react": "^6.0.2",
     "@meronex/icons": "^4.0.0",
-    "@netlify/functions": "^0.11.0",
-    "@netlify/plugin-nextjs": "^5.10.0",
-    "@next/bundle-analyzer": "^12.1.0",
-    "@notionhq/client": "^2.3.0",
-    "@types/jsdom": "^20.0.1",
-    "@types/react-sticky-el": "^1.0.3",
-    "@types/react-syntax-highlighter": "^13.5.2",
-    "@types/rss": "^0.0.29",
-    "@types/uuid": "^8.3.4",
-    "axios": "^0.25.0",
-    "date-fns": "^2.28.0",
-    "date-fns-tz": "^1.2.2",
-    "dotenv": "^16.5.0",
+    "@netlify/functions": "^5.1.2",
+    "@netlify/plugin-nextjs": "^5.15.3",
+    "@next/bundle-analyzer": "^16.1.1",
+    "@notionhq/client": "^5.6.0",
+    "@types/jsdom": "^27.0.0",
+    "@types/react-sticky-el": "^2.1.0",
+    "@types/react-syntax-highlighter": "^15.5.13",
+    "@types/rss": "^0.0.32",
+    "@types/uuid": "^11.0.0",
+    "axios": "^1.13.2",
+    "date-fns": "^4.1.0",
+    "date-fns-tz": "^3.2.0",
+    "dotenv": "^17.2.3",
     "encoding": "^0.1.13",
-    "esbuild": "^0.20.2",
-    "framer-motion": "^6.5.1",
-    "fs": "^0.0.1-security",
-    "jsdom": "^20.0.3",
+    "esbuild": "^0.27.2",
+    "framer-motion": "^12.23.26",
+    "fs": "0.0.1-security",
+    "jsdom": "^27.4.0",
     "lodash": "^4.17.21",
     "net": "^1.0.2",
-    "netlify-cli": "^12.4.1",
-    "next": "^15.2.2",
+    "netlify-cli": "^23.13.0",
+    "next": "^16.1.1",
     "next-absolute-url": "^1.2.2",
-    "next-seo": "^5.15.0",
-    "next-sitemap": "^3.1.43",
-    "node-loader": "^2.0.0",
-    "react": "^18.3.1",
-    "react-code-blocks": "^0.0.9-0",
-    "react-dom": "^18.3.1",
-    "react-icons": "^4.3.1",
-    "react-intersection-observer": "^9.16.0",
+    "next-seo": "^7.0.1",
+    "next-sitemap": "^4.2.3",
+    "node-fetch": "^3.3.2",
+    "node-loader": "^2.1.0",
+    "react": "^19.2.3",
+    "react-code-blocks": "^0.1.6",
+    "react-dom": "^19.2.3",
+    "react-icons": "^5.5.0",
+    "react-intersection-observer": "^10.0.0",
     "react-share": "^5.2.2",
-    "react-sticky-el": "^2.0.9",
-    "react-syntax-highlighter": "^15.5.0",
-    "remark": "^14.0.1",
-    "remark-html": "^15.0.0",
-    "skia-canvas": "^2.0.2",
-    "swr": "^1.2.1",
+    "react-sticky-el": "^2.1.1",
+    "react-syntax-highlighter": "^16.1.0",
+    "remark": "^15.0.1",
+    "remark-html": "^16.0.1",
+    "skia-canvas": "^3.0.8",
+    "swr": "^2.3.8",
     "tls": "^0.0.1",
-    "uuid": "^8.3.2"
+    "uuid": "^13.0.0"
   },
   "devDependencies": {
     "@netlify/esbuild": "^0.14.39",
-    "@testing-library/jest-dom": "^6.6.3",
-    "@testing-library/react": "^16.3.0",
-    "@types/jest": "^29.5.14",
-    "@types/node": "^22.13.10",
-    "@types/react": "^18.3.18",
-    "@typescript-eslint/parser": "^5.12.1",
-    "eslint-import-resolver-typescript": "^2.5.0",
-    "eslint-plugin-import": "^2.25.4",
-    "feed": "^4.2.2",
+    "@testing-library/jest-dom": "^6.9.1",
+    "@testing-library/react": "^16.3.1",
+    "@types/jest": "^30.0.0",
+    "@types/node": "^25.0.3",
+    "@types/react": "^19.2.7",
+    "@typescript-eslint/parser": "^8.51.0",
+    "eslint-import-resolver-typescript": "^4.4.4",
+    "eslint-plugin-import": "^2.32.0",
+    "feed": "^5.1.0",
     "identity-obj-proxy": "^3.0.0",
-    "jest": "^29.7.0",
-    "ts-jest": "^29.3.2",
-    "typescript": "^5.8.2"
+    "jest": "^30.2.0",
+    "ts-jest": "^29.4.6",
+    "typescript": "^5.9.3"
   }
 }
