+++ renovate.json
@@ -1,0 +1,15 @@
+{
+  "$schema": "https://docs.renovatebot.com/renovate-schema.json",
+  "extends": [
+    "config:recommended"
+  ],
+  "packageRules": [
+    {
+      "matchUpdateTypes": ["major", "minor", "patch"],
+      "stabilityDays": 1,
+      "description": "Wait 1 day after release before updating"
+    }
+  ],
+  "timezone": "Asia/Tokyo",
+  "schedule": ["after 9am and before 5pm on weekdays"]
+}
