+++ src/pages/tags/[tag]/page/[page].tsx
@@ -1,19 +1,9 @@
 import React from 'react';
-import dynamic from 'next/dynamic';
 import { GetStaticPropsContext, GetStaticPathsResult } from 'next';
 import { ParsedUrlQuery } from 'querystring';
-import { ACTIVE_THEME } from '../../../../config/themeSelector';
 import { IPageHead } from 'core/types/NotionPageApiResponses';
 import { CommonDataService } from '../../../../services/CommonDataService';
-
-// 動的にテーマのホームページコンポーネントをインポート
-const HomePage = dynamic(() =>
-  import(`../../../../themes/${ACTIVE_THEME}/pages/HomePage`).then(mod => mod.default),
-  {
-    loading: () => null,
-    ssr: true,
-  }
-);
+import HomePage from '../../../../themes/theme2/pages/HomePage';
 
 // 1ページあたりの記事数
 const POSTS_PER_PAGE = 10;
